add HSK1 extra words; #17

+++ vocab/autogen-hsk.rb
@@ -4,12 +4,18 @@ require 'erb'
 require 'set'
 
 RUST_FILE = "../src/autogen_hsk.rs"
-HSK1 = {src: "hsk1.tsv", qc: "hsk1-QC-do-not-edit.tsv"}
-HSK2 = {src: "hsk2.tsv", qc: "hsk2-QC-do-not-edit.tsv"}
+WORD_FILES = [
+  "hsk1.tsv",
+  "hsk1-extra.tsv",
+  "hsk2.tsv",
+]
 
 # Returns array: [[hanzi, pinyin], [hanzi, pinyin], ...]
+# The `select {..."#"..."\t"}` filters out blank lines and comments
 def read_tsv(file)
-  File.read(file).lines.map { |n| n.chomp.split("\t") }
+  File.read(file).lines
+    .select { |n| !n.start_with?("#") && n.include?("\t") }
+    .map { |n| n.chomp.split("\t") }
 end
 
 # Make a set with each unique character used in pinyin from <file>.
@@ -27,26 +33,16 @@ def normalize(pinyin)
   return n
 end
 
-# Check integrity and coverage of the character transposition table
-detected = (char_set(HSK1[:src]) + char_set(HSK2[:src])).to_a.sort.join("")
+# Check integrity and coverage of the character transposition table.
+# The map/reduce uses set algebra to build a sorted string of unique characters from all the files.
+detected = WORD_FILES.map {|wf| char_set(wf)}.reduce {|a,b| a+b}.to_a.sort.join("")
 if detected != TR_FROM
-  warn "Error: Characters used in pinyin of #{HSK1[:src]} or #{HSK2[:src]} do not match TR_FROM"
+  warn "Error: Characters used in word file pinyin do not match TR_FROM"
   warn " detected: \"#{detected}\""
   warn " TR_FROM:  \"#{TR_FROM}\""
   abort "You need to update TR_FROM and TR_TO so pinyin will properly normalized to ASCII"
 end
 abort "Error: Check for TR_FROM/TR_TO length mismatch" if TR_FROM.size != TR_TO.size
-
-# Generate a quality check TSV file for manually checking the normalized pinyin
-print "This will overwrite #{HSK1[:qc]}, #{HSK2[:qc]}, and #{RUST_FILE}\nProceed? [y/N]: "
-abort "no changes made" if !["y", "Y"].include? gets.chomp
-for h in [HSK1, HSK2]
-  File.open(h[:qc], "w") { |qc|
-    for hanzi, pinyin in read_tsv(h[:src])
-      qc.puts "#{hanzi}\t#{pinyin}\t#{normalize(pinyin)}"
-    end
-  }
-end
 
 # Merge hanzi values for duplicate pinyin search keys
 # example: ["he", "he"] and ["喝", "和"] get turned into ["he"] and ["喝\t和"]
@@ -55,18 +51,22 @@ merged_pinyin = []
 first_index_of = {}
 duplicate_pinyin = []
 i = 0
-for level in [HSK1, HSK2]
-  for hanzi, pinyin in read_tsv(level[:src])
-    nrmlzd_pinyin = normalize(pinyin)
-    if first_index_of[nrmlzd_pinyin]
+for wf in WORD_FILES
+  for hanzi, pinyin in read_tsv(wf)
+    normalized_pinyin = normalize(pinyin)
+    if first_index_of[normalized_pinyin]
+      if merged_hanzi[first_index_of[normalized_pinyin]].include?(hanzi)
+        # If you see this warning, check for two files listing the exact same word
+        warn "Likely duplicate word: #{wf}:  #{hanzi}:#{pinyin}  ==>  try:  grep #{hanzi} *.tsv"
+      end
       # Duplicate search key ==> Append hanzi to first entry
-      merged_hanzi[first_index_of[nrmlzd_pinyin]] += "\t#{hanzi}"
-      duplicate_pinyin << nrmlzd_pinyin
+      merged_hanzi[first_index_of[normalized_pinyin]] += "\t#{hanzi}"
+      duplicate_pinyin << normalized_pinyin
     else
       # First instance of search key ==> Add new entries
       merged_hanzi[i] = hanzi
-      merged_pinyin[i] = nrmlzd_pinyin
-      first_index_of[nrmlzd_pinyin] = i
+      merged_pinyin[i] = normalized_pinyin
+      first_index_of[normalized_pinyin] = i
       i += 1
     end
   end
@@ -74,6 +74,11 @@ end
 
 # Sort the merged vocab lists in pinyin order
 merged_pinyin, merged_hanzi = merged_pinyin.zip(merged_hanzi).sort.transpose
+duplicate_pinyin = duplicate_pinyin.sort.uniq
+
+puts "\nPreparing to generate rust source code..."
+print "This will overwrite #{RUST_FILE}\nDo you want to continue? [y/N] "
+abort "no changes made" if !["y", "Y"].include? gets.chomp
 
 # Generate rust source code with hanzi and pinyin arrays
 File.open(RUST_FILE, "w") { |rf|
@@ -82,7 +87,7 @@ File.open(RUST_FILE, "w") { |rf|
     // To make changes, see ../vocab/autogen-hsk.rb
 
     // The hanzi values for these duplicate pinyin search keys were merged:
-    <% duplicate_pinyin.uniq.each do |dp| %>//  <%= dp %>
+    <% duplicate_pinyin.each do |dp| %>//  <%= dp %>
     <% end %>
     pub const HANZI: &[&'static str] = &[
     <% merged_hanzi.each do |h| %>    &"<%= h %>",
